Multi-platform build system. Working on linux. Probably broken on osx for g++ specific flags.

+++ src/kernel/parser/parser.sh
@@ -6,16 +6,18 @@ set -e
 # turn on for verbose logging
 # set -xv
 
-# OSX HACK: hardcoded for homebrew gcc49 package
-if [ -z "$CC" ]; then CC=/usr/local/bin/g++-4.9; fi
-MAGLEV_HOME=$(cd $(dirname $0)/../../.. ; pwd)
-GEMSTONE=$MAGLEV_HOME/gemstone
-# HACK
-GSVERSION=3.1.0.2.1-64
-# HACK
-EXT=dylib
+OS=$(uname -s)
+
+case $OS in
+    "Linux" | "Darwin" ) ;; # good
+    "*"                ) echo "Unknown OS: ${OS}"; exit 1 ;;
+esac
 
 cd $(dirname $0)
+
+MAGLEV_HOME=$(cd ../../.. ; pwd)
+GEMSTONE=$MAGLEV_HOME/gemstone
+GSVERSION=3.1.0.2.1-64                       # HACK
 
 ./yacc.sh
 
@@ -23,23 +25,31 @@ if test $? -ne 0; then
   echo "  byacc failed"
   exit 1
 fi
-chmod -w rubygrammar.{h,c}
 echo "  byacc ok"
 
-echo "Compiling rubygrammar.o"
+if [ -z "$CC" ]; then
+    case $OS in
+        "Linux" ) CC=/usr/bin/g++; ;;
+        "Darwin") CC=/usr/local/bin/g++-4.9 ;;
+    esac
+fi
 
-CCWARN="-Wchar-subscripts -Wcomment -Werror -Wformat -Wmissing-braces -Wmultichar -Wno-aggregate-return -Wno-c++11-extensions -Wno-constant-logical-operand -Wno-conversion -Wno-deprecated-declarations -Wno-format -Wno-invalid-offsetof -Wno-self-assign-field -Wno-unused-function -Wparentheses -Wreturn-type -Wshadow -Wsign-compare -Wsign-promo -Wswitch -Wsystem-headers -Wtrigraphs -Wunused-label -Wunused-value -Wunused-variable -Wwrite-strings"
+case $OS in
+    "Linux" )
+        EXT=so
+        ;;
+    "Darwin")
+        EXT=dylib
+        ;;
+esac
 
-# TODO: most of these look like legacy copy&paste.
-CCDEF="-DFLG_DEBUG=1 -D_XOPEN_SOURCE -D_REENTRANT -DNOT_JAVA_VM -D_GNU_SOURCE -D_LARGEFILE64_SOURCE -Dlint"
+# TODO: probably need to push this into the above case for OS/compiler specifics
+CCWARN="-Wchar-subscripts -Wcomment -Werror -Wformat -Wmissing-braces -Wmultichar -Wno-aggregate-return -Wno-unused-function -Wparentheses -Wreturn-type -Wshadow -Wsign-compare -Wsign-promo -Wswitch -Wsystem-headers -Wtrigraphs -Wtrigraphs -Wuninitialized -Wunused-label -Wunused-value -Wunused-variable -Wwrite-strings"
+CCDEF="-DFLG_FAST=1 -DNOT_JAVA_VM -D_GNU_SOURCE -D_REENTRANT"
+CCINC="-I$GEMSTONE/include -I."
+CFLAGS="$CCDEF $CCINC $CCWARN -O3 -fPIC -fcheck-new -fmessage-length=0 -fno-exceptions -fno-strict-aliasing -g -m64 -pipe -pthread -x c++"
 
-LDFLAGS="-Wl,-flat_namespace -Wl,-undefined -Wl,warning -Wl,-exported_symbols_list -Wl,magparse.osx.exp"
-
-$CC $CCWARN $CCDEF \
-  -g -m64 -pipe -pthread -fmessage-length=0 -fPIC -fno-strict-aliasing -fno-exceptions \
-  -I. -I$GEMSTONE/include -I/usr/local/Cellar/icu4c/54.1/include \
-  -x c++ \
-  -c rubygrammar.c -o rubygrammar.o
+$CC $CFLAGS -c rubygrammar.c -o rubygrammar.o
 
 if test $? -ne 0; then
   echo "compiling rubygrammar.o failed"
@@ -47,11 +57,7 @@ if test $? -ne 0; then
 fi
 
 echo "Compiling rubyast.o"
-$CC $CCWARN $CCDEF \
-  -g -m64 -pipe -pthread -fmessage-length=0 -fPIC -fno-strict-aliasing -fno-exceptions \
-  -I. -I$GEMSTONE/include -I/usr/local/Cellar/icu4c/54.1/include \
-  -x c++ \
-  -c rubyast.c -o rubyast.o
+$CC $CFLAGS -c rubyast.c -o rubyast.o
 
 if test $? -ne 0; then
   echo "compiling rubyast.o failed"
@@ -60,17 +66,19 @@ fi
 
 rm -f libmagparse.${EXT}
 
-echo "Linking libmagparse.so"
-$CC \
-  -m64 -lpthread -ldl -lc -lm  -o libmagparse.${EXT} \
-  $LDFLAGS \
-  -Wl,-dylib \
-  -L$GEMSTONE/lib \
-  rubyast.o rubygrammar.o
+echo "Linking libmagparse.${EXT}"
+LDFLAGS="-shared -Wl,--version-script=magparse.exp -Wl,--warn-unresolved-symbols -Wl,-Bdynamic,-hlibmagparse.${EXT} -lc -lcrypt -ldl -lm -lpthread -lrt -m64"
+
+$CC $LDFLAGS rubyast.o rubygrammar.o -o libmagparse.${EXT}
+
+if test $? -ne 0; then
+  echo "linking libmagparse.${EXT} failed"
+  exit 1
+fi
 
 chmod 555 libmagparse.${EXT}
 
 echo "Copying libmagparse.${EXT} to $GEMSTONE/lib/"
 chmod +w $GEMSTONE/lib
 rm -f $GEMSTONE/lib/libmagparse-$GSVERSION.${EXT}
-cp libmagparse.${EXT} $GEMSTONE/lib/libmagparse-3.1.0.2.1-64.${EXT}
+cp libmagparse.${EXT} $GEMSTONE/lib/libmagparse-$GSVERSION.${EXT}
